refactor(platform/host): move symbols for configuring liveness probes to `client.ts`
