+++ projects/scion/microfrontend-platform/src/lib/host/client-registry/client.ts
@@ -69,3 +69,21 @@ export interface Client {
    */
   dispose(): void;
 }
+
+/**
+ * Specifies the interval (in milliseconds) at which the host pings connected clients.
+ *
+ * By default, if not set, a ping interval of 60s is used.
+ *
+ * @internal
+ */
+export const CLIENT_PING_INTERVAL = Symbol('CLIENT_PING_INTERVAL');
+
+/**
+ * Specifies the ping timeout (in milliseconds) for unregistering a client.
+ *
+ * By default, if not set, clients are unregistered if not answering the ping within 10 seconds.
+ *
+ * @internal
+ */
+export const CLIENT_PING_TIMEOUT = Symbol('CLIENT_PING_TIMEOUT');
