+++ projects/scion/microfrontend-platform/src/lib/host/client-registry/ɵclient.ts
@@ -15,8 +15,7 @@ import {MessageClient} from '../../client/messaging/message-client';
 import {PlatformTopics} from '../../ɵmessaging.model';
 import {Logger, LoggingContext} from '../../logger';
 import {ClientRegistry} from './client.registry';
-import {CLIENT_PING_INTERVAL, CLIENT_PING_TIMEOUT} from './client.constants';
-import {Client} from './client';
+import {Client, CLIENT_PING_INTERVAL, CLIENT_PING_TIMEOUT} from './client';
 import {ɵApplication, ɵVERSION} from '../../ɵplatform.model';
 import {IntentSubscription, IntentSubscriptionRegistry} from '../message-broker/intent-subscription.registry';
 import {UUID} from '@scion/toolkit/uuid';
